impact-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
strawberry-graphql[fastapi]==0.217.1
motor==3.6.0
beanie==1.26.0
httpx==0.26.0
pydantic-settings==2.1.0
